地图追踪接口修改成多实例，JS调用的实例和地图追踪调用的实例不同
