+++ BetterGenshinImpact/Core/Script/Dependence/Genshin.cs
@@ -42,6 +42,12 @@ public class Genshin
     /// 系统屏幕的DPI缩放比例
     /// </summary>
     public double ScreenDpiScale => TaskContext.Instance().DpiScale;
+    
+    public Lazy<NavigationInstance> LazyNavigationInstance { get; } = new(() =>
+    {
+        Navigation.WarmUp();
+        return new NavigationInstance();
+    });
 
     /// <summary>
     /// 传送到指定位置
@@ -214,7 +220,7 @@ public class Genshin
             throw new InvalidOperationException("不在主界面，无法识别小地图坐标");
         }
 
-        return MapManager.GetMap(mapName).ConvertImageCoordinatesToGenshinMapCoordinates(Navigation.GetPositionStable(imageRegion, mapName));
+        return MapManager.GetMap(mapName).ConvertImageCoordinatesToGenshinMapCoordinates(LazyNavigationInstance.Value.GetPositionStable(imageRegion, mapName));
     }
 
     #endregion 大地图操作
